+++ SpaceZD.API/Models/Outputs/CarriageTypeOutputModel.cs
@@ -5,5 +5,5 @@ public class CarriageTypeOutputModel
     public int Id { get; set; }
     public string Name { get; set; }
     public int NumberOfSeats { get; set; }
-    public double PriceFactor { get; set; }
+    public decimal PriceFactor { get; set; }
 }
